bind/Test,capps-bind: Now can pass args to named; always use --rm
All arguments to `capps-bind` previously were passed to docker; now if
you add a `--` argument the arguments after that will be passed to the
container process instead. A side effect of this implementation is
that empty string args can no longer be passed to docker; this is
probably not an issue.

We also now always use --rm because I don't want to be accidentally
leaving a lot of containers around. (We can still add --rm=false if we
want to override this.) We can revisit this later if necessary, but
ensuring the logs are outside the container and the container itself
is storing no useful non-transient state seems like a good way to go.