bind/capps-bind: Mount the config dir in the container
We use /etc/bind by default, but a different dir can be supplied with
the -C option. The dir must be absolute because of the way the `docker
-v` option works; this is checked before trying to continue.

Because bind drops most root privileges, the config dir and its
contents must be readable to the `bind` user. Remember that when you
bind mount files into a Docker container they keep the same owners and
permissions as they have on the host.