bind/capps-bind: Add logging
We bind-mount /dev/log into the container so that processes in it log
just as if they were not running in a container, and now tell named to
log to stderr instead of syslog only in test mode (-T).

We also add the -L option to send stdout/stderr to the systemd journal
so that it's preserved even after the container exits. (The default
log driver is usually json-file, which stores the output with the
container, deleting it when the container is deleted.) `docker logs`
still works to view it as well, as long as the container still exists.