dent/Test: Add --no-force-rebuild to Test
This avoids rebuilding layers that are already cached, which can be
useful when testing parts of the build functionality outside of those
that configure the Docker container. It can also be used to more
quickly test only later layers of the build if you manually remove
other otherwise invalidate (e.g. by changing Dockerfile) those layers.