dent/: Alternative for pwent on Windows
On Windows we get the username through alternate, more portable means
and just make up a UID (1234) and a GECOS (full name, etc.) field.
This could probably be improved a bit.

This now lets `py dent` compile on Windows and display the help
message; it hasn't really been tested further than that. In
particular, the shebang (when run with `./dent`) doesn't work because
it requests Python 3; we need to look for a nice way of with this in
Windows.