dent/: Context files explicitly UTF-8; don't set perms on Windows
XXX Not Python 2 compatible, and not sure how to deal with this. Binary mode?

At least one of the files we write out for the Docker context already
has non-ASCII characters in it, and this can fail on any system where
the user's locale is not using UTF-8. So in Python 3 we should always
be explicitly stating the encoding. This may even be true of the the
program file itself; that requires further investigation.

Also, on Windows we must explicitly set the newline format for the
`setup-image` script not to use CRs because Bash will interpret those
as extra characters at the end of the line and complain when these
disrupt the syntax.

Windows has no fchmod call to set the permissions of the files in the
context, but it doesn't really make sense to do this anyway, at least
not with a chmod-like call, because on Windows that can only change
the read-only setting. Docker seems to use some reasonable default for
the file permissions if we don't touch them, so let's just not touch
them for the moment and see if any problems arise.