dent/: Explicitly specifiy UTF-8 when writing files
If the the system or interpreter default encoding is not UTF-8,
writing `setup-image` may fail because it has non-ASCII characters. We
always want to use UTF-8, since we're not concerned about interactions
with other things, just what goes on in our container.

This shows up as a test failure only in certain environments, but it's
rather difficult to set up tests to reliably identify this problem.
Since it's fairly obvious and easily fixed when it fails, anyway, we
just live without proper tests for it.