dent/dent: Fix git.{user,email} setting
This didn't work properly becuase we tried to run git to set
git.{user,email} before git was installed. So we now install git
first, do the setting, and then install etckeeper. (On Ubuntu the
etckeeper package installation also does the initial commit, but we
continue to do an `etckeeper init` after because Debian does not do
this on installation.)