dent/dent: Add --dry-run option
This is a bit more subtle than it seems, since many of the Docker commands
we run still need to be executed in a dry run so that we're making the
right decisions about what image and container creation/removal/etc.
commands we would be running.

A new set of tests (well, one test) has been added for the dry run; this
also happens to add some (slightly fragile) testing that correct parameters
are being passed to `docker run` and so on. It's not clear how useful this
is, and how much of the testing with actual runs this might be able to
replace.

Reviewed-by: Nishant Rodrigues <nishantjr@gmail.com>