dent/dent: Flush stdout/err before exec
Unlike calling exit() or generating a SystemExit exception, no interpreter
cleanup is done before an os.execvp() so any unflushed data is lost. (The
documentation[1] explains this; but we apparently did not read that.)

This was causing some very mysterious problems in new tests we were writing
that used `out=$(...)` since some output did appear; it was just stdout
that was vanishing, and only when captured by $(...).

Fixing this also surfaced a test where we weren't using -q, but the test
was passing anyway (at least on some systems) beacuse that output was
instead being lost.

[1]: https://docs.python.org/3/library/os.html#os.execvp

Reviewed-by: Nishant Rodrigues <nishantjr@gmail.com>