dent/dent: With --keep-tmpdir, print tmpdir location even if -q specified
If we suppressed that message with -q and --keep-tmpdir, the directory
would not be removed but we'd have no indication that stuff was left
behind, much less where it was left.

Unfortunately there's no test for this; it would need to use --keep-tmpdir
and then find and remove the directory afterwards (ideally handling being
interrupted as well), which doesn't seem worth all the effort.

Reviewed-by: Nishant Rodrigues <nishantjr@gmail.com>