dent: .gitignore {hosts,resolv.conf} so etckeeper will install
Possibly due to changes in newer versions of Docker, /etc/hosts and
/etc/resolv.conf are imported from the host so that changes in the host
networking are reflected in the container. (I thought it had always done
this, but apparently not; this is breaking even on the debian:8 image.)

This now breaks something during `apt-get install`, where something happens
that that seems to convince etckeeper that it needs to reset the
permissions on those files:

    55.15 Processing triggers for libc-bin (2.28-10) ...
    55.72 chmod: changing permissions of 'hosts': Read-only file system

Adding `/hosts/` and `/resolv.conf` to /etc/.gitignore before attempting
to install etckeeper fixes this.