dent/Test: Fix calculated image name so it's removed properly
Dent's name for the image it creates from the base image was changed a
while back: for a `debian:10` base image it used to use `dent-debian-10`
but now uses `dent/debian.10`. However, the test code was not change,
so it was trying to clean up images under the old name instead of the new,
and so left them behind. This is now fixed.

We also slightly tweak the name of the test container to better match the
test image, though there was no bug here since the test code supplies the
test container name to dent.