dent: Improve argument parsing with add_mutually_exclusive_group()
Previously we still had to specify a container even for the
-L/--list-base-images option, though that doesn't use a container
name. We fix this by creating a mutually exclusive group for -L and
the container name so that one or the other, but not both, must be
provided.

We also use this technique for the two mutually exclusive arguments
that help specify an image name, -i and -t, replacing the custom code
with which we did this before.

Additionally, explain why we use nargs=REMAINDER instead of nargs='*',
which we'd looked at as another way of dealing with this. (It's not
only not as good a way, but inconvenient for the end user since
it brings back the need often to use `--` in the command line.