bin/dent: Add commented-out --progress=plain for build command
We need to add an option to dent to enable this, since the new builds show
only the last few lines of output and thus can miss error messages.

A question on stackoverflow.com[1] asked how to view these logs, and the
only answers mention --progress=plain, so there appears to be no log of the
build.

[1]: https://stackoverflow.com/q/37832575/107294