bin/dent SETUP_IMAGE dot_home(): Don't run apt-get!
The dot_home() function was for some reason runing apt-get to install
curl, git and sudo. This of course breaks Fedora etc., and isn't
necessary at all because all of these packages are already part of
$UNIVERSAL_PKGS which are installed by the distro-specific install
functions. So remove this line completely.

This does raise the point that perhaps not only should packages() be
in a separate layer from users() and dot_home(), to speed rebuilds of
fresh dent containers for other users or whatever, but we might
consider doing two packages() layers, as the first one will get most
of the updates (between the parent image release and the time of
build), letting you remove the second to get a much faster set of
differental package updates if you re-build later to get newer
packages. Or perhaps those package updates should be part of the
user()/dot_home() image.