dent: Set $USER and $LOGNAME in container
Since we are trying to emulate a user environment, we must have the
$LOGNAME (required by POSIX) and $USER (not required by POSIX, but
extremely common) variables set. These are normally set by login(1),
but since we don't run that to enter the container, we must set them
ourselves.

`Test` has been updated to confirm that these variables are set. Note
that it explicitly does not use the `-l argument to bash because these
variables should be available even when the user's runnning a one-off
command like `dent deb11 ls`.

This also avoids false negatives (i.e., incorrectly passing) from
systems like Fedora that have an /etc/profile that sets and exports
LOGNAME and USER. (Debian and Ubuntu do not do this, which is how
we discovered the problem.)