dent: Add optional pre-setup command to run before /tmp/setup
Alpine Linux doesn't have `bash` installed by default, so the attempt
to run `/bin/bash /tmp/setup-image` fails.

The second argument of BASE_IMAGES has always been an empty dict; we
now use this as a configuration dictionary. If it contains a
'presetup' key, we take the value as a command to give (unquoted) to
`RUN` in the Dockerfile before we run /tmp/setup-image, so we can
install bash or whatever.

If no 'presetup' command is given, it defaults to `true`.

This doesn't handle the situation where someone specifies an image not
in the BASE_IMAGES list but similar enough to need the pre-setup,
e.g., 'alpine:3.18' instead of 'alpine:3.19'. Possibly we could deal
with this by searching for a matching prefix, or possibly we should
just have a separate config for base image names without version
numbers that apply to all versions.

Note that this commit does not yet make Alpine Linux run: further work
on SETUP_IMAGE is needed for that.