dent: Add alpine:3.19 and support for alternate user creation routine
Alpine Linux appears not to have the `useradd` command available,
so instead we must use `adduser`.

`adduser` is also available on Debian (though it's not a default package),
but Alpine's `adduser` isn't compatible (e.g., it wants `-g` instead of
`--coment`), so it's easier just to have a separate user setup function.