dent: Split setup-image into setup-pkg and setup-user
This splits the package setup (which takes quite a while) and the user
setup (which is quick) into two separate layers, so that rebuilding a
container is _much_ faster when only user information changes.

There's still no easy way to find and remove that layer when we want just
to re-build the user info (e.g., the dot-home setup); we should consider
splitting the Docker image setup into two parts to add an explicit name to
this if we want to do that. Or perhaps we could just put a the current
date/time or something into the SETUP_USER script.