dent: Fix "tries to build container it should download" bug
This was discovered by running these tests on a system that did not have an
`alpine:latest` image available locally. Rather than downloading it, dent
attempted to build its own version of that image `FROM None`. (Oops!)

What we now do is check to see if we have the requested image (in which
case we don't build) OR we were given an explicit image name with the
--image NAME option, in which case we assume it's already built, either
locally or available via download. Thus --image NAME now works the same as
just passing NAME to `docker run`.

This bumps the patch version number.