Test: Fix error collection bugs and improve display
Apparently we didn't have everything using adderror(); some tests were
still were still just incrementing $retcode, which has been ignored since
adderror() was implemented. Fix those.

We also improve the error output: adderror() prints the error itsef when
called, so we can go back through the output stream and see the output
around where it happened, and the $errorlist array is also printed again at
the end so that it's easy to see all the errors that occurred. (Calls to
adderror() now prefix the base image so we know in which test the failure
happened.) This also now lets us just exit with code 1 instead of a code
equal to the number of errors, which was kind of pointless since the number
of errors is now printed.