bin/dent: Set verifypw=any for the user so sudo -v never needs a pw
We previously didn't add the user to the sudo or wheel groups because
that added (on some systems anyway, e.g. Debian) sudo entries that
did not include NOPASSWD: and with the default `verifypw = all` setting
that meant that `sudo -v` required a password.

However, we can fix that by setting `verifypw = any` for the user,
and then it doesn't matter what other groups she's in or about
anything else that somehow gets added to her sudoers configuration.