dent: Add HOST_HOSTNAME env var to containers
This will allow scripts in containers to know the name of the host on
which the container is running. It's mainly intended for things like
`xtitle` or prompts that display the hostname, so they can say it's
not just a container called 'deb12', but also show the name of the
host running that container. (This is particularly helpful when you're
logged in to a remote machine and using a dent container there.)