*: Convert this to a proper Python distribution package
The major version number has been bumped from 0 to 1 not because there's
been any functionality change (if anything, it's not clear if this change
might not introduce bugs) but because the way this is installed has been
vastly changed, and it's _very_ helpful to make that clear when doing any
kind of debugging. Further, this makes it clear that we are well
superseding the empty 0.1.0 package for the previous project that used
[dent] on PyPI.

The current version is 1.0.0.dev1 to accommodate a few more changes (see
below) and a test upload to PyPI before we officially release 1.0.0.beta1.

The `description` line has been well reduced from GitHub because the GitHub
version is too complicated. Note that this must be a single line (not clear
on any length limitations) as it will be shown below the package name in a
PyPI search.

Notes:
- The version number is curently duplicated in `dent` and `pyproject.toml`;
  this will be fixed in a later commit.
- doc/release.md is copied from [r8format]. It's not ideal to have this
  duplicated; we need to work out some way of consolidating our release
  process documentation and scripting into one place.

[r8format]: https://github.com/mc68-net/r8format
[dent]: https://pypi.org/project/Dent/