WIP *: Convert this to a proper Python distribution package
The major version number has been bumped from 0 to 1 not because there's
been any functionality change (if anything, it's not clear if this change
might not introduce bugs) but because the way this is installed has been
vastly changed, and it's _very_ helpful to make that clear when doing
any kind of debugging.

The `description` line has been well reduced from GitHub because the GitHub
version is too complicated. Note that this must be a single line (not clear
on any length limitations) as it will be shown below the package name in a
PyPI search.

XXX Test : bin/dent → dent throughout

TODO:
- XXX version number duplicated in `dent` and `pyproject.toml` figure out
  how to fix this
- XXX fix r8format's tool.setuptools.package.py-package crap; should not be
  there...
- CHANGELOG!
- Document how we do releases (ref to r8format).
- Update docs for new install and give examples of pae/pipx/etc. use

- XXX explain why no:

        [tool.setuptools.packages]
        find = { where = ['src'] }

        [tool.setuptools]
        py-modules = ['pytest_pt']