dockerd-proxy: Add setup.cfg to install this because we have to
Note that this is not updated automatically with an editable install (`pip
install -e`); you need actually to reinstall the package to update it.

bin/dockerd-proxy cannot be installed using configuration in pyproject.toml
because setuptools does not support this. Instead, we must specify in
`setup.cfg` that this file should be copied into the installed bin/
directory, and this will not be updated from the source even when the
package is installed as editable. (See the comments in `setup.cfg` for more
information.)

Trying to do this from pyproject.toml:

    [tool.setuptools.options]
    scripts = 'bin/dockerd-proxy'

We get this error:

    ValueError: invalid pyproject.toml config: `tool.setuptools`.
    configuration error: `tool.setuptools` must not contain {'options'} properties