dent: PROGNAME should be based on argv[0], not __file__
In the original "all in one file" version of dent, __file__ is always
the program itself. But as we move towards becoming a proper Python
distribution package, __file__ becomes `dent.py` or something else,
depending on in which source PROGNAME is set.

So instead we use os.basename(sys.argv[0]), which will still be the
file itself in the "all in one file version," but will be set to the
nominal name of the file we're running (i.e., still `…/dent`) by the
shim that runs main().

This issue was discovered when we noticed in a testing version
that the image names had changed from e.g. `dent/debian.12` to
`dent.py/debian.12`.

(No version number bump on this because it doesn't change the
current behaviour at all.)