src/dent: Get version number via importlib.metadata
This lets us have the version number in only one place, the pyproject.toml.
`dent --version` now reads the version number from the metadata generated
for the package.

The importlib.metadata interface requires at least version 3.8, so we no
longer support Python 3.6 and 3.7, but those are quite old so this
shouldn't be a big issue.

Note that even in an editable install the package version is updated only
on an actual package installation, so if you change the version number and
want to see the new one you'll have to re-install dent.