build-release: Script to automate part of the package release process
The instructions in `doc/release.md` are updated for this.

`build` and `twine` are removed from `requirements.txt` as the
`build-release` script takes care of installing these. It's better
that it does that because these are moderately large packages that
most developers don't ever need, so there's no reason to make all
developers install them with every clean build.