Test,etc.: Add pytest unit test framework; this changes `Test` args
We now have a simple test in src/dent/main.pt that mainly confirms
that the unit test framework is working. Further tests will require
some restructuring of main.py, particularly in how global variables
are used.

In order to pass arguments to pytest from Test we need to rework the
command line arguments. It now uses -B IMGNAME options to determine
which images the "build" section of the tests should run (instead of
these being straight command-line args), and stops parsing after -- so
that options as well as command-line args can be passed to pytest.

Note that the -v option for Test does not affect pytest; for that use
`./Test -- -v`.

Also clean up requirements.txt slightly: pytest-pt will bring in
pytest for us, and there's no need to specify a minumum version since
we're not likely to "accidentally" get a lower version of pytest than
the latest.