*: Add mypy type checking
We lock the checking to Python version 3.8 (which limit works regardless of
what version of Python you're actually using) as that will help us maintain
compatibility with intended minimum version for the project, also 3.8. (It
would be nice if there were a way to just set this to
$project.project-version.)

This found a lot of errors in `dent.main`, so we've disabled the checking
for those particular errors for the moment until we fix them in a later
commit. (Some of them may just go away with our reworking of the vars in
that file anyway.)

Test now accepts a `-T` option to skip the type check. (And the existing
`-v` option will make mypy verbose, though I'm not sure if we really
want that.)