dent.main: Fix all legacy type errors
These were pretty easy to fix, and also exposed a global variable that
we weren't "declaring."

We use e.g. `typing.Dict[str,str]` instead of `dict[str,str]` because
the latter doesn't become availble until 3.9, and it's not worth
dropping 3.8 compatibility just for this. (3.8 is still supported,
even if only for a couple more months, and it remains in Ubuntu 20.04
supported until 2025, with security releases until 2030.)