dent.config: Take over PWENT and ARGS globals
We now pass a Config into all those functions that used to be using
these globals, as the first step to not having globals to be able to
more easily unit test functions.

There are still a couple of globals (PROGNAME and IMAGE_CONF) that
need to be cleaned up.