src/dent/tmpl/: Extract Dockerfile/script templates to separate files
src/dent/tmpl/ now contains separate files (Dockerfile, setup_head.bash,
setup_packages.bash and setup_user.bash) with the templates we use
to create files to set up the image.

This requires upgrading our minimum Python version from 3.8 to 3.9, as the
importlib.resources module is not available yet in 3.8. (This means we no
longer work on Debian 10, Ubuntu 20.04, and probaly CentOS 8.)

We could instead use pkg_resources, but this is inefficient and
discouraged, and no doubt will be going away at some point. We could also
use the [backport], but it doesn't seem worth the effort. (That said, we
can always add that and bump the version back down again if we get
complaints.)

Now that we're at 3.9, we should go through and update the typing stuff at
some point to take advantage of no longer needed to import so much stuff
from `typing`.

[backport]: https://importlib-resources.readthedocs.io/