dent.main: Change shell comments to fix "command not found" messages
The command to put text in a file under sudoers.d/ had backticked
`all` and `any` in comments, but because this was being read via
`<<` the shell actually tried to execute these and substitute before
putting the comment into the file. Change these to 'all' and 'any'
(with single quotes rather than backticks) to avoid the error
messages and get the right text into the file.