dent.main SETUP_USER: Remove UID 1000 if it exists
Linux users using this program often have a UID of 1000. However,
the Ubuntu images include an `ubuntu` user with UID 1000, which
collides with the UID of the dent user when she's gone with the
default UID that Debian, Ubuntu and many other distros assign.

So remove the user with UID 1000 if it's present in the image
before we create the user that's a copy of the user running dent.