Extract embedded files from dent.py and load with importlib_resources
This now requires PyPI packages `importlib_resources` (because that
provides a consistent interface across all versions of Python[1]) and
`setuptools-scm` (to include the resource files in the built package).

That the files are included in built packages has been tested only
manually by building the wheel, listing its contents, and installing
and running it in a virtualenv. This test is not _that_ difficult to
automate, but now that we've confirmed it working manually, it seems
unlikely to fail.

[1]: We use the PyPI importlib_resources compatibility package because
the importlib.resources API changed enough from Python 3.7 to 3.11
that it's a pain to use and test across all versions. For further info
on importlib.resources and importlib_resources, see [2].

[2]: https://github.com/cynic-net/py-resource-test