main docker_inspect(): Clean up and type this
We change the args to take just a single container name, since the
ability to use multiple container names was never used and just made
things a bit more confusing. We now can also change the return type
to be just a dict (or None), so that callers need not index into
the always 0- or 1-length list.